+++ src/AdaptiveLighting.Lamplight/GlobalUsings.cs
@@ -1,0 +1,10 @@
+global using System;
+global using System.Collections.Generic;
+global using System.Linq;
+global using System.Threading.Tasks;
+
+global using Microsoft.Extensions.Logging;
+
+global using AdaptiveLighting.Hosting;
+global using AdaptiveLighting.Web.Presentation;
+global using AdaptiveLighting.Web.Services;